moved d64NiceGui to own module
Improved the project structure for development of further tools and moved first tool to its own module.
